refactor(core): split pipeline logic into decide and merge
