+++ src/merge.ts
@@ -1,14 +1,9 @@
 import debug from 'debug';
 import _ from 'lodash';
+import { getMergeDecision, updateDecisionsForDependsOn } from './decide';
 import { ARTIFACT_INJECTION_STEP_KEY, artifactInjectionSteps, nothingToDoSteps } from './steps';
-import { count } from './util';
 
-const log = debug('monofo:pipeline');
-
-const decide = (included: boolean, reason: string): IncludeDecision => ({
-  included,
-  reason,
-});
+const log = debug('monofo:merge');
 
 /**
  * Loop through the decided configurations and, for any excluded parts, collect the keys of steps that are now skipped.
@@ -44,47 +39,6 @@ function replaceExcludedKeys(configs: ConfigWithDecision[], hasArtifactStep: boo
   });
 }
 
-export interface IncludeDecision {
-  included: boolean;
-  reason: string;
-}
-
-/**
- * If a config has changes, its steps are merged into the final build. Otherwise, it is excluded, and its excluded_steps
- * are merged in instead. There are exceptions:
- *  - An env var named PIPELINE_RUN_ALL, set to any value, indicates that all steps should run
- *  - An env var named PIPELINE_RUN_<NAME>, where NAME is the UPPER_SNAKE_CASE version of the component pipeline name,
- *    set to any value, indicates that step should run
- */
-function getMergeDecision(config: ConfigWithChanges): IncludeDecision {
-  if (process.env.PIPELINE_RUN_ALL) {
-    return decide(true, 'been forced to by PIPELINE_RUN_ALL');
-  }
-
-  const envVarName = config.monorepo.name.toLocaleUpperCase().replace('-', '_');
-
-  const overrideExcludeKey = `PIPELINE_NO_RUN_${envVarName}`;
-  if (process.env[overrideExcludeKey]) {
-    return decide(false, `been forced NOT to by ${overrideExcludeKey}`);
-  }
-
-  const overrideIncludeKey = `PIPELINE_RUN_${envVarName}`;
-  if (process.env[overrideIncludeKey]) {
-    return decide(true, `been forced to by ${overrideIncludeKey}`);
-  }
-
-  if (!config.buildId) {
-    return decide(true, 'no previous successful build, fallback');
-  }
-
-  // At this point, we know it has a config.buildId we can grab artifacts from
-  if (config.changes.length > 0) {
-    return decide(true, `${count(config.changes, 'matching change')}: ${config.changes.join(', ')}`);
-  }
-
-  return decide(false, 'no matching changes');
-}
-
 function toMerge({ steps, env, included, monorepo }: ConfigWithDecision): Pipeline {
   return included
     ? {
@@ -102,29 +56,9 @@ function toPipeline(steps: Step[]): Pipeline {
 }
 
 /**
- * Mutates the config objects within ConfigWithDecision to account for transitive dependencies between pipelines
- */
-function updateDecisionsForDependsOn(configs: ConfigWithDecision[]): void {
-  const byName = Object.fromEntries(configs.map((c) => [c.monorepo.name, c]));
-
-  configs
-    .flatMap((config) => config.monorepo.depends_on.map((dependency) => [config.monorepo.name, dependency]))
-    .reverse()
-    .forEach(([from, to]) => {
-      const dependent = byName[from];
-      const dependency = byName[to];
-
-      if (dependent.included && !dependency.included) {
-        dependency.included = true;
-        dependency.reason = `been pulled in by a depends_on from ${from}`;
-      }
-    });
-}
-
-/**
  * @param results
  */
-export function mergePipelines(results: ConfigWithChanges[]): Pipeline {
+export default function mergePipelines(results: ConfigWithChanges[]): Pipeline {
   log(`Merging ${results.length} pipelines`);
 
   const decisions: ConfigWithDecision[] = results.map((r) => {
